getting rid of useless 'TABLET_*_COUNTERS' - keeping only executor counters; moved TMetrics to tablet_counters.h/cpp
